backend/requirements.txt
sentence-transformers
faiss-cpu
langchain
langchain-community
pandas